Fix build on Linux.
Math functions (pow, floor, etc.) require a separate math library on Linux (probably some other platforms too.  That wasn't being linked so the final linkage didn't work on Linux when library was used as port of an exe or built as a shared lib.

This change adds a platform check for a math function in the math library and links it if found.  CMake usage requirements should take care of linking it into the final binary.